Moved DonationForm component code into its own folder within components and changed DonationForm's CSS file to CSS modules
