+++ package.json
@@ -31,8 +31,6 @@
     "eject": "react-scripts eject",
     "lint": "eslint '**/*.{js,jsx}'",
     "format": "prettier --write \"**/*.+(js|jsx|css)\"",
-    "storybook": "start-storybook -p 6006 -s public",
-    "build-storybook": "build-storybook -s public",
     "prepare": "husky install"
   },
   "lint-staged": {
@@ -67,12 +65,6 @@
     ]
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^6.3.12",
-    "@storybook/addon-essentials": "^6.3.12",
-    "@storybook/addon-links": "^6.3.12",
-    "@storybook/node-logger": "^6.3.12",
-    "@storybook/preset-create-react-app": "^4.1.2",
-    "@storybook/react": "^6.3.12",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.25.2",
