+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { ChakraProvider } from '@chakra-ui/react';
 import React from 'react';
 import './App.css';
-import DonationForm from './components/DonationForm';
+import DonationForm from './components/DonationForm/DonationForm';
 
 function App() {
   return (
